patch(update) Fix Drupal.attachBehaviors after update
Apparently now you need to pass the context for attaching behaviors and it doesn't just work like it used to. Drupal...amirite? Also, there needs to be an attached contextual link on the page you started out in otherwise you are screwed because the behaviour never existed for you to attach...again Drupal AMARITE?